Disallow Server config file download.
Laravel comes with pre-defined server configuration files (.htaccess for apache servers and web.config for IIS). Because those files are located in the mounted `public` directory, it is currently possible to download them by visiting the URI with the file path (www.example.com/.htaccess).

This change is to explicitly disallow access to .htaccess and .config files.